% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ap_dx.R
\docType{data}
\name{ap_dx}
\alias{ap_dx}
\title{APACHE Diagnosis Map Dataset}
\format{An R data.frame}
\usage{
data(ap_dx)
}
\description{
A mapping table used to map different kinds of admission diagnoses in a hiearchical structure.
}
\keyword{datasets}
