% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{generate_formula}
\alias{generate_formula}
\title{Generate mgcv model forumula to fit the Global Open Source Severity of Illness Scale.}
\usage{
generate_formula(basis, kay)
}
\arguments{
\item{basis}{"\"cr\"" or "\"cr\"" string to be used in fitting}

\item{kay}{the degrees of freedom for the fit}
}
\value{
a string that can be used as a formula.
}
\description{
This function generates a the model formula to be fit with gam/bam
}
\details{
Helper function to generate a model formula.
Used mainly in tuning the model during development.
}
\examples{
require(GOSSIS)
# new_dat <- preprocess_data(dat)
# imp_dat <- impute_data(dat,algo=3)
# dat <- prepare_fit(imp_dat)
# here
# fit  <- gam(formula(formm),data=dp_io2,family="binomial",na.action=na.exclude)

}
\author{
Jesse D. Raffa
}
