% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gossis1_ihmp.R
\docType{data}
\name{gossis1_ihmp}
\alias{gossis1_ihmp}
\title{GOSSIS prediction model datasets}
\format{An R data.frame}
\usage{
data(gossis1_ihmp)
}
\description{
A `gam` model with some elements stripped  used in GOSSIS in-hospital mortality prediction.
}
\seealso{
gpredict
}
\keyword{datasets}
