% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_models.R
\docType{data}
\name{imputation_models}
\alias{imputation_models}
\title{GOSSIS imputation model datasets}
\format{An R data.frame}
\usage{
data(imputation_models)
}
\description{
A named list of imputation models used for GOSSIS.
}
\seealso{
imputate_data
}
\keyword{datasets}
