% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_models_fac.R
\docType{data}
\name{imputation_models_fac}
\alias{imputation_models_fac}
\title{GOSSIS imputation model datasets (factors variables only).}
\format{An R data.frame}
\usage{
data(imputation_models)
}
\description{
A named list of imputation models used for GOSSIS (factors variables only).
}
\seealso{
imputate_data
}
\keyword{datasets}
