% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{impute_data}
\alias{impute_data}
\title{Apply imputation procedure for the Global Open Source Severity of Illness Scale.}
\usage{
impute_data(x, algo = 3, files = NULL, median_Overide = FALSE,
  mediandata = NULL)
}
\arguments{
\item{x}{a data.frame with the minimal GOSSIS-1 dataset, and correct column names, that has been preprocessed with preprocess_data}

\item{algo}{the imputation algorithm selected to use.  defaults to option 3.
Option 1 is a model where all predictors (except the variable to be imputed and the dx_sub variable) are used in prediction.
This algorithm is applied to all missing data.
Option 2 is a model which uses all predictors except the variable to be imputed, the variables related to the variable being imputated (e.g., the other _min/_max and _apache variables of the same variable).
The algorithm is applied to all missing data.
Option 3 is a hybrid of algorithm 1 and 2, applying algorithm 1 to those observations which have a similar variable (e.g., someone with _apache, but missing _max), and algorithm 2 for those missing all of _min, _max and _apache variables of the same type.
Option 4 imputes the median or mode of each variable using the training dataset.}

\item{files}{optional argument to specify the files which to use for imputation.  By default it uses the ones used in the paper.}
}
\value{
a data.frame that has been completed
}
\description{
This function imputes missing data based on models developed for the Global Open Source Severity of Illness Scale.
It offers several variants of imputation, including median imputation, with the suggested implementation, algorithm 3
}
\details{
The function applies a series of prediction models used for imputation of missing data contained in a GOSSIS dataset.
There are currently 4 variants of the algorithm implemented.  All four return a completed (missing data filled in) dataset.
Please use this function carefully.  It is not recommended for non-GOSSIS applications.
}
\examples{
require(GOSSIS)
# new_dat <- preprocess_data(dat)
# imp_dat <- impute_data(dat,algo=3)

}
\author{
Jesse D. Raffa
}
