% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medians_numeric.R
\docType{data}
\name{medians_numeric}
\alias{medians_numeric}
\title{GOSSIS imputation model median imputation method's median values to impute numeric missing data.}
\format{An R data.frame}
\usage{
data(medians_numeric)
}
\description{
A one row data.frame with the median values of the training set which was used to perform median imputation on the GOSSIS algorithm for numeric data.
Note: This method is not used, it was implemented for comparison only.
}
\seealso{
imputate_data
}
\keyword{datasets}
