% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modes_cat.R
\docType{data}
\name{modes_cat}
\alias{modes_cat}
\title{GOSSIS imputation model mode imputation method's mode values to impute}
\format{An R data.frame}
\usage{
data(medians_numeric)
}
\description{
A one row data.frame with the mode values of the training set which was used to perform mode imputation on the GOSSIS algorithm for categorical variables.
Note: This method is not used, it was implemented for comparison only.
}
\seealso{
imputate_data
}
\keyword{datasets}
