% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prepare_fit}
\alias{prepare_fit}
\title{Prepare data frame for model fitting in the Global Open Source Severity of Illness Scale.}
\usage{
prepare_fit(x)
}
\arguments{
\item{x}{a data.frame with the minimal GOSSIS-1 dataset, and correct column names, that has been preprocessed with preprocess_data}
}
\value{
a data.frame that has been prepared
}
\description{
This function prepares the data for fitting the prediction models
}
\details{
The function applies a series of prediction models used for imputation of missing data contained in a GOSSIS dataset.
There are currently 4 variants of the algorithm implemented.  All four return a completed (missing data filled in) dataset.
Please use this function carefully.  It is not recommended for non-GOSSIS applications.
}
\examples{
require(GOSSIS)
# new_dat <- preprocess_data(dat)
# imp_dat <- impute_data(dat,algo=3)
# dat <- prepare_fit(imp_dat)

}
\author{
Jesse D. Raffa
}
