% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Apply simple preprocessing of a data.frame in preparation for GOSSIS imputation}
\usage{
preprocess_data(dat)
}
\arguments{
\item{dat}{a data.frame with the minimal GOSSIS-1 dataset, and correct column names.}
}
\value{
a data.frame that has been processed and ready to begin imputation.
}
\description{
This function applies some simple preprocessing to a data.frame that is to undergo imputation.
Most operations inolve binning categories or handling missing diagnosis data.
A full description of transformations can be found in the code comments and below.
}
\details{
The function joins the ap_dx data.frame to dat, mapping the diagnoses into group, dx_class, and dx_sub.
It also handles missing or mismatched diagnoses.
The function also maps the 3 GCS score components into a single 4 level factor variable.
}
\examples{
require(GOSSIS)
# new_dat <- preprocess_data(dat)

}
\seealso{
ap_dx
}
\author{
Jesse D. Raffa
}
